Pack de Atualização com adição de Chat, e controles administrativos de arquivos e Sistema

+++ backend/src/utils/deleteFile.ts
@@ -1,13 +1,16 @@
 import fs from 'fs';
 import path from 'path';
 
-// Função para deletar um arquivo do sistema
+// Função para deletar um arquivo do sistema, se ele existir
 export function deleteFile(filePath: string) {
   const fullPath = path.resolve(__dirname, '..', '..', 'uploads', filePath);
-  
-  fs.unlink(fullPath, (err) => {
-    if (err) {
-      console.error(`Erro ao tentar deletar o arquivo ${filePath}:`, err);
-    }
-  });
+
+  // Verifica se o arquivo existe antes de tentar deletá-lo
+  if (fs.existsSync(fullPath)) {
+    fs.unlink(fullPath, (err) => {
+      if (err) {
+        console.error(`Erro ao tentar deletar o arquivo ${filePath}:`, err);
+      }
+    });
+  }
 }
